microservices/data-access-service/src/main.py
import logging
import os
from dotenv import load_dotenv
import uvicorn
from .api import app

load_dotenv()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000) 