microservices/data-access-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
boto3==1.34.34
pyarrow==15.0.0
pandas==2.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==8.0.0
httpx==0.26.0
structlog==24.1.0
prometheus-client==0.19.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0 